+++ app/src/main/java/dev/kichan/marketplace/model/repository/CouponRepositoryImpl.kt
@@ -1,32 +1,13 @@
-package dev.kichan.marketplace.common
+package dev.kichan.marketplace.model.repository
 
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import dev.kichan.marketplace.model.NetworkModule
-import dev.kichan.marketplace.model.data.CouponMemberRes
 import dev.kichan.marketplace.model.data.LatestCoupon
 import dev.kichan.marketplace.model.data.ResponseTemplate
+import dev.kichan.marketplace.model.service.CouponPageNation
 import dev.kichan.marketplace.model.service.CouponService
-import kotlinx.coroutines.launch
 import retrofit2.Response
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
-
-class CouponViewModel : ViewModel() {
-    private val repository = CouponRepositoryImpl()
-
-    val coupons = MutableLiveData<List<CouponMemberRes>>()
-    fun getCoupons(marketId: Long, couponId: Long? = null, size: Int? = null) {
-        viewModelScope.launch {
-            val res = repository.getCoupons(marketId, couponId, size)
-
-            if(res.isSuccessful) {
-                coupons.value = res.body()!!.response.couponResDtos
-            }
-        }
-    }
-}
 
 class CouponRepositoryImpl {
     private val service = NetworkModule.getService(CouponService::class.java)
@@ -38,7 +19,7 @@ class CouponRepositoryImpl {
         lastCreatedAt: LocalDateTime?,
         lastCouponId: Long?,
         pageSize: Int?
-    ): Response<ResponseTemplate<List<LatestCoupon>>> {
+    ): Response<ResponseTemplate<CouponPageNation>> {
         return service.getLatestCoupon(
             if (lastCreatedAt != null) {
                 val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS")
