+++ app/src/main/java/dev/kichan/marketplace/ui/page/MyPage.kt
@@ -3,15 +3,12 @@ package dev.kichan.marketplace.ui.page
 import LargeCategory
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.icons.outlined.Info
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -33,7 +30,7 @@ import androidx.navigation.compose.rememberNavController
 import dev.kichan.marketplace.R
 import dev.kichan.marketplace.ui.Page
 import dev.kichan.marketplace.ui.bottomNavItem
-import dev.kichan.marketplace.ui.component.dev.kichan.marketplace.AuthViewModel
+import dev.kichan.marketplace.AuthViewModel
 import dev.kichan.marketplace.ui.component.dev.kichan.marketplace.ui.component.atoms.BottomNavigationBar
 import dev.kichan.marketplace.ui.component.dev.kichan.marketplace.ui.component.atoms.CategorySelector
 import dev.kichan.marketplace.ui.component.dev.kichan.marketplace.ui.component.atoms.MyPageCard
@@ -43,7 +40,7 @@ import dev.kichan.marketplace.ui.theme.PretendardFamily
 @Composable
 fun MyPage(navController: NavController, viewModel : AuthViewModel) {
     val member = viewModel.member.observeAsState()
-    val myCuration = viewModel.myCuration.observeAsState()
+//    val myCuration = viewModel.myCuration.observeAsState()
 
     var selectedCategory by remember {
         mutableStateOf(mutableListOf(LargeCategory.All))
@@ -169,31 +166,31 @@ fun MyPage(navController: NavController, viewModel : AuthViewModel) {
             Spacer(modifier = Modifier.height(20.dp))
 
             // MyPageCard를 세로로 나열하는 리스트, 각 카드 사이에 구분선 추가
-            if(myCuration.value.isNullOrEmpty()) {
-                Box(
-                    modifier = Modifier.fillMaxWidth(),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text("저장한 매장이 없습니다.", modifier = Modifier.padding(vertical = 16.dp))
-                }
-            }
-            else {
-                LazyColumn(
-                    contentPadding = PaddingValues(horizontal = 16.dp, vertical = 8.dp),
-                    verticalArrangement = Arrangement.spacedBy(0.dp) // 카드 사이 간격 제거
-                ) {
-                    items(myCuration.value ?: listOf()) { event ->
-                        Column {
-                            MyPageCard(event = event)
-                            Divider(
-                                color = Color(0xFFF4F4F4),
-                                thickness = 1.dp,
-                                modifier = Modifier.fillMaxWidth()
-                            )
-                        }
-                    }
-                }
-            }
+//            if(myCuration.value.isNullOrEmpty()) {
+//                Box(
+//                    modifier = Modifier.fillMaxWidth(),
+//                    contentAlignment = Alignment.Center
+//                ) {
+//                    Text("저장한 매장이 없습니다.", modifier = Modifier.padding(vertical = 16.dp))
+//                }
+//            }
+//            else {
+//                LazyColumn(
+//                    contentPadding = PaddingValues(horizontal = 16.dp, vertical = 8.dp),
+//                    verticalArrangement = Arrangement.spacedBy(0.dp) // 카드 사이 간격 제거
+//                ) {
+//                    items(myCuration.value ?: listOf()) { event ->
+//                        Column {
+//                            MyPageCard(event = event)
+//                            Divider(
+//                                color = Color(0xFFF4F4F4),
+//                                thickness = 1.dp,
+//                                modifier = Modifier.fillMaxWidth()
+//                            )
+//                        }
+//                    }
+//                }
+//            }
         }
     }
 }
